pyproject.toml
[project]
name = "kage"
version = "0.1.0"
description = "KAGE - Krigsexe Agentic Generative Engine - Local AI coding assistant"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Julien GELEE", email = "krigsexe@proton.me" }
]
keywords = ["ai", "agent", "llm", "coding-assistant", "mcp", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    # CLI
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.0",
    
    # LLM
    "ollama>=0.3.0",
    "openai>=1.40.0",
    "tiktoken>=0.7.0",
    
    # Memory & Storage
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    
    # Tools
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "gitpython>=3.1.0",
    "pyyaml>=6.0.0",
    "toml>=0.10.0",
    
    # Validation & Config
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    
    # Utils
    "structlog>=24.4.0",
    "tenacity>=8.5.0",
    "watchfiles>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
kage = "kage.cli.app:main"

[project.urls]
Homepage = "https://github.com/Krigsexe/Kage"
Repository = "https://github.com/Krigsexe/Kage"
Issues = "https://github.com/Krigsexe/Kage/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kage"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PT", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/kage --cov-report=term-missing"
